requirements.txt
fastapi
pydantic
google-generativeai
uvicorn
pydantic